pyproject.toml
[tool.poetry]
name = "rag_medical_poc"
version = "0.1.0"
description = "Healthcare RAG system for automated ICD-10-CM/CPT code lookup with hybrid retrieval"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
openai = "^1.50.0"
langchain-openai = "^0.2.1"
langchain-community = "^0.3.0"
langchain = "^0.3.0"
faiss-cpu = "^1.9.0"
spacy = "~=3.7.5"
scispacy = "^0.5.4"
scikit-learn = "^1.5.2"
sentence-transformers = "^3.1.0"
rank-bm25 = "^0.2.2"
ragas = "^0.1.19"
streamlit = "^1.39.0"
structlog = "^24.4.0"
pydantic = "^2.9.0"
pandas = "^2.2.3"
numpy = "^1.26.0"
openpyxl = "^3.1.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
ruff = "^0.7.0"
black = "^24.10.0"
jupyter = "^1.1.1"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=85"
